pyproject.toml
[project]
name = "briefmarker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.9.0",
    "flask>=3.0.3",
    "flask-cors>=5.0.0",
    "inference>=0.22.2",
    "opencv-python-headless>=4.10.0.84",
    "python-dotenv>=1.0.1",
]
